Ensure build artifacts are cleaned out
Not sure why "ruby-${RUBY_VERSION}"{,.tar.gz} didn't work here, but it
clearly didn't, which added 300M of uncached useless fat to the image.

Added a test to ensure the image stays under 600M (it currently should
be ~500M), which should be a high watermark and an indicator that
something went wrong.